move nextflow utils
